Refactor - Redux toolkit을 사용하여 action creators, reducers를 slice.js로 구현
