+++ package.json
@@ -42,6 +42,7 @@
   },
   "dependencies": {
     "@codeceptjs/configure": "^0.5.2",
+    "@reduxjs/toolkit": "^1.4.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
